+++ liquid/future/tags/__init__.py
@@ -1,8 +1,10 @@
 from ._case_tag import LaxCaseTag  # noqa: D104
 from ._if_tag import LaxIfTag
+from ._tablerow_tag import InterruptingTablerowTag
 from ._unless_tag import LaxUnlessTag
 
 __all__ = (
+    "InterruptingTablerowTag",
     "LaxCaseTag",
     "LaxIfTag",
     "LaxUnlessTag",